Split and cleanup test files
